refactor: 정보 함수 모듈 분리
