+++ lib/kim/calendarModal.js
@@ -1,5 +1,5 @@
 import { getUserBookmarkList } from "./userBookmark.js";
-import { getCompanies } from "./renderCalendar.js";
+import { getCompanies } from "./getInfo.js";
 import { currentUser } from "../data/currentUser.js";
 import dummyUsers from "../data/dummyUserData.js";
 import dummyJobPostings from "../data/dummyJobPostingsData.js";
