+++ lib/kim/renderCalendar.js
@@ -1,10 +1,11 @@
 import { getUserBookmarkList } from "./userBookmark.js";
+import { getCompanies, getDays } from "./getInfo.js";
 import { currentUser } from "../data/currentUser.js";
 import dummyUsers from "../data/dummyUserData.js";
 import dummyJobPostings from "../data/dummyJobPostingsData.js";
 
-let currentYear = new Date().getFullYear();
-let currentMonth = new Date().getMonth() + 1;
+export let currentYear = new Date().getFullYear();
+export let currentMonth = new Date().getMonth() + 1;
 
 const prevMonthBtn = document.querySelector("#prev-month-btn");
 const nextMonthBtn = document.querySelector("#next-month-btn");
@@ -50,35 +51,14 @@ function renderDay(currentYear, currentMonth){
 
     const days = getDays(currentYear, currentMonth);
     
+    // 유저 북마크 리스트의 회사명, 마감기한 전체
     const userbookmarkedInfo = getUserBookmarkList(currentUser, dummyUsers(), dummyJobPostings());
 
     for(let day = 1; day <= days; day++){
+        // 날짜 별 유저 북마크 회사명
         const companies = getCompanies(day, userbookmarkedInfo);
         calendarContainer.insertAdjacentHTML('beforeend', createDay(day, companies));
     }
-}
-
-// 유저, 날짜별 채용 공고 정보 가져오는 함수
-export function getCompanies(day ,info){
-    const companies = [];
-    for(const job of info){
-        const endDate = new Date(job.endDate);
-
-        if (
-        endDate.getFullYear() === currentYear &&
-        endDate.getMonth() + 1 === currentMonth &&
-        endDate.getDate() === day
-        ) {
-        companies.push(job.company);
-      }
-    }
-    return companies;
-}
-
-// 그 달의 날짜 가져오는 함수
-export function getDays(currentYear, currentMonth){
-    const days = new Date(currentYear, currentMonth, 0).getDate();
-    return days;
 }
 
 // 날짜 요소 만드는 함수
