+++ lib/kim/userBookmark.js
@@ -1,4 +1,4 @@
-// 현재  user 북마크 리스트의 회사명, 마감기한 가져오는 함수
+// 현재  user 북마크 리스트의 회사명, 마감기한 리스트 가져오는 함수
 export function getUserBookmarkList(currentUser, userbookmark, jobList){
   // user목록에서 uesr 찾기
     const user = userbookmark.find((user) => {
